requirements.txt
statsmodels
corner
emcee
mpmath
numpy
scipy
pandas
matplotlib
rasterio
cartopy
pillow
